carga banco Primera, con resultado de carga, falta mostrar la validacion y colorear filas con error

+++ 01_DataLayer/documentosRecibidos.cs
@@ -43,9 +43,9 @@ namespace DataLayer
 			dAccess.AddParameter(ref pPos, "@TipoDocumentoID", objDocumentosRecibidos.TipoDocumentoID, SqlDbType.BigInt, 0, 0, ParameterDirection.Input, ref Param);
 			dAccess.AddParameter(ref pPos, "@NaturalezaAdquisicion", objDocumentosRecibidos.NaturalezaAdquisicion, SqlDbType.VarChar, 0, 0, ParameterDirection.Input, ref Param);
 			dAccess.AddParameter(ref pPos, "@NumeroDocumentoCausa", objDocumentosRecibidos.NumeroDocumentoCausa, SqlDbType.VarChar, 0, 0, ParameterDirection.Input, ref Param);
-			dAccess.AddParameter(ref pPos, "@ValorNeto", objDocumentosRecibidos.ValorNeto, SqlDbType.BigInt, 18, 3, ParameterDirection.Input, ref Param);
-			dAccess.AddParameter(ref pPos, "@ValorIVAFactura", objDocumentosRecibidos.ValorIVAFactura, SqlDbType.BigInt, 18, 3, ParameterDirection.Input, ref Param);
-			dAccess.AddParameter(ref pPos, "@ValorTotalFactura", objDocumentosRecibidos.ValorTotalFactura, SqlDbType.BigInt, 18, 3, ParameterDirection.Input, ref Param);
+			dAccess.AddParameter(ref pPos, "@ValorNeto", objDocumentosRecibidos.ValorNeto, SqlDbType.Float, 18, 3, ParameterDirection.Input, ref Param);
+			dAccess.AddParameter(ref pPos, "@ValorIVAFactura", objDocumentosRecibidos.ValorIVAFactura, SqlDbType.Float, 18, 3, ParameterDirection.Input, ref Param);
+			dAccess.AddParameter(ref pPos, "@ValorTotalFactura", objDocumentosRecibidos.ValorTotalFactura, SqlDbType.Float, 18, 3, ParameterDirection.Input, ref Param);
 			dAccess.AddParameter(ref pPos, "@LugarDocumento", objDocumentosRecibidos.LugarDocumento, SqlDbType.VarChar, 0, 0, ParameterDirection.Input, ref Param);
 			dAccess.AddParameter(ref pPos, "@FechaDocumento", objDocumentosRecibidos.FechaDocumento, SqlDbType.VarChar, 0, 0, ParameterDirection.Input, ref Param);
 			dAccess.AddParameter(ref pPos, "@NombreAutorizanteEmisor", objDocumentosRecibidos.NombreAutorizanteEmisor, SqlDbType.VarChar, 0, 0, ParameterDirection.Input, ref Param);
